Pins read and set, can move with the map on first floor.
Read the showcases from the database and save them as empty cases in the mainActivity, and pass them to the firstfloor and secondfloor to draw the edges and the Pins. And the pins for the first floor can move with the walls now, but the scale is still need to be implemented.